+++ metadrive/utils/waymo_utils/filter_cases.py
@@ -8,9 +8,8 @@ from metadrive.envs.real_data_envs.waymo_env import WaymoEnv
 from metadrive.policy.idm_policy import WaymoIDMPolicy
 
 try:
-    from metadrive.utils.waymo_utils.waymo_utils import AgentType
-    from metadrive.utils.waymo_utils.waymo_utils import RoadEdgeType
-    from metadrive.utils.waymo_utils.waymo_utils import RoadLineType
+    from metadrive.utils.waymo_utils.waymo_type import AgentType
+    from metadrive.utils.waymo_utils import RoadLineType, RoadEdgeType
 finally:
     pass
