+++ metadrive/utils/waymo_utils/waymo_type.py
@@ -1,0 +1,40 @@
+LaneType = {
+    0: 'UNKNOWN',
+    1: 'LANE_FREEWAY',
+    2: 'LANE_SURFACE_STREET',
+    3: 'LANE_BIKE_LANE'
+}
+RoadLineType = {
+    0: 'UNKNOWN',
+    1: 'ROAD_LINE_BROKEN_SINGLE_WHITE',
+    2: 'ROAD_LINE_SOLID_SINGLE_WHITE',
+    3: 'ROAD_LINE_DOUBLE_WHITE',
+    4: 'ROAD_LINE_BROKEN_SINGLE_YELLOW',
+    5: 'ROAD_LINE_BROKEN_DOUBLE_YELLOW',
+    6: 'ROAD_LINE_SOLID_SINGLE_YELLOW',
+    7: 'ROAD_LINE_SOLID_DOUBLE_YELLOW',
+    8: 'ROAD_LINE_PASSING_DOUBLE_YELLOW'
+}
+RoadEdgeType = {
+    0: 'UNKNOWN',
+    1: 'ROAD_EDGE_BOUNDARY',
+    2: 'ROAD_EDGE_MEDIAN'
+}
+AgentType = {
+    0: 'UNSET',
+    1: 'VEHICLE',
+    2: 'PEDESTRIAN',
+    3: 'CYCLIST',
+    4: 'OTHER'
+}
+TrafficSignal = {
+    0: 'LANE_STATE_UNKNOWN',
+    1: 'LANE_STATE_ARROW_STOP',
+    2: 'LANE_STATE_ARROW_CAUTION',
+    3: 'LANE_STATE_ARROW_GO',
+    4: 'LANE_STATE_STOP',
+    5: 'LANE_STATE_CAUTION',
+    6: 'LANE_STATE_GO',
+    7: 'LANE_STATE_FLASHING_STOP',
+    8: 'LANE_STATE_FLASHING_CAUTION'
+}
